Move create mask to train_utils
